Update and delete transaction

+++ entity/transaction.go
@@ -1,7 +1,7 @@
 package entity
 
 type Transaction struct {
-	ID               int    `gorm:"primary_key:auto_increment" json:"id"`
+	ID               uint64 `gorm:"primary_key:auto_increment" json:"id"`
 	UserID           uint64 `gorm:"index" json:"user_id"`
 	TransactionType  string `json:"transaction_type"`
 	Date             string `json:"date"`
